server/requirements.txt
# Web framework
flask>=3.0.0
flask-cors>=4.0.0

# Tableau Server Client
tableauserverclient>=0.31

# Environment variables
python-dotenv>=1.0.0

# HTTP requests (for Trello in Phase 4)
requests>=2.31.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0

# Development
black>=24.0.0
